requirements.txt
opencv-python==4.6.0.66
gps==3.9
ultralytics==8.2.46
mysql-connector-python==8.4.0
folium==0.15.1

